// lib/types/notification-events.ts
export enum NotificationEvent {
  // API Key Events
  API_KEY_CREATED = 'API_KEY_CREATED',
  API_KEY_UPDATED = 'API_KEY_UPDATED',
  API_KEY_DELETED = 'API_KEY_DELETED',
  API_KEY_EXPIRING = 'API_KEY_EXPIRING',
  API_KEY_EXPIRED = 'API_KEY_EXPIRED',
  
  // User Events
  USER_CREATED = 'USER_CREATED',
  USER_UPDATED = 'USER_UPDATED',
  USER_DELETED = 'USER_DELETED',
  USER_ROLE_CHANGED = 'USER_ROLE_CHANGED',
  USER_JOINED_ORG = 'USER_JOINED_ORG',
  
  // Organization Events
  ORG_CREATED = 'ORG_CREATED',
  ORG_UPDATED = 'ORG_UPDATED',
  ORG_SUBSCRIPTION_CHANGED = 'ORG_SUBSCRIPTION_CHANGED',
  ORG_LIMIT_WARNING = 'ORG_LIMIT_WARNING',
  ORG_LIMIT_EXCEEDED = 'ORG_LIMIT_EXCEEDED',
  
  // Cost & Usage Events
  COST_THRESHOLD_WARNING = 'COST_THRESHOLD_WARNING',
  COST_THRESHOLD_CRITICAL = 'COST_THRESHOLD_CRITICAL',
  COST_THRESHOLD_EXCEEDED = 'COST_THRESHOLD_EXCEEDED',
  DAILY_COST_SPIKE = 'DAILY_COST_SPIKE',
  UNUSUAL_SPENDING_PATTERN = 'UNUSUAL_SPENDING_PATTERN',
  
  // API Testing Events
  API_TEST_STARTED = 'API_TEST_STARTED',
  API_TEST_COMPLETED = 'API_TEST_COMPLETED',
  API_TEST_FAILED = 'API_TEST_FAILED',
  API_PROVIDER_OUTAGE = 'API_PROVIDER_OUTAGE',
  API_RATE_LIMIT_WARNING = 'API_RATE_LIMIT_WARNING',
  API_RATE_LIMIT_EXCEEDED = 'API_RATE_LIMIT_EXCEEDED',
  
  // Thread Events
  THREAD_SHARED = 'THREAD_SHARED',
  THREAD_COLLABORATION_INVITE = 'THREAD_COLLABORATION_INVITE',
  THREAD_MESSAGE_RECEIVED = 'THREAD_MESSAGE_RECEIVED',
  
  // System Events
  SYSTEM_UPDATE = 'SYSTEM_UPDATE',
  MAINTENANCE_SCHEDULED = 'MAINTENANCE_SCHEDULED',
  FEATURE_ANNOUNCEMENT = 'FEATURE_ANNOUNCEMENT',
  
  // Alert Events  
  ALERT_TRIGGERED = 'ALERT_TRIGGERED',
  ALERT_RESOLVED = 'ALERT_RESOLVED',
  ALERT_ACKNOWLEDGED = 'ALERT_ACKNOWLEDGED',
  
  // Site-wide Announcements
  SITE_ANNOUNCEMENT = 'SITE_ANNOUNCEMENT',
  CRITICAL_ALERT = 'CRITICAL_ALERT',
  PROMOTIONAL_OFFER = 'PROMOTIONAL_OFFER'
}